requirements.txt
crewai==0.1.24
requests
python-dotenv